refactor: 拆分 agent.rs 为模块目录，提升可维护性
将 2940 行的 agent.rs 拆分为 agent/ 模块目录，按职责分离：
- types.rs: 核心类型定义
- config.rs: 常量与配置工具
- content.rs: 多模态内容构建器
- events.rs: 前端事件发射
- api_types.rs: SSE/API DTO
- context.rs: 上下文管理
- errors.rs: 错误处理
- providers/: 四种 Provider 独立实现
公共 API 保持不变，外部调用方无需修改。
