+++ src-tauri/src/agent/providers/mod.rs
@@ -1,0 +1,4 @@
+pub(super) mod anthropic;
+pub(super) mod openai_chat;
+pub(super) mod openai_responses;
+pub(super) mod codex;
